fix(core): use the quint64 QMetaEnum overload in the capability round-trip
valueToKey(int) is QT_CORE_REMOVED_SINCE(6,9), so the int cast resolved
to the quint64 overload through an implicit conversion that trips
-Wsign-conversion on a cold (non-ccached) configure. Cast to quint64
explicitly in the manager and the test.

+++ src/core/update/update_manager.cpp
@@ -32,8 +32,9 @@ QUrl UpdateManager::feedUrl() const {
 }
 
 QString UpdateManager::capabilityToString(Capability capability) {
+    // valueToKey(int) is QT_CORE_REMOVED_SINCE(6,9); only the quint64 overload is visible.
     return QString::fromLatin1(
-        QMetaEnum::fromType<Capability>().valueToKey(static_cast<int>(capability)));
+        QMetaEnum::fromType<Capability>().valueToKey(static_cast<quint64>(capability)));
 }
 
 UpdateManager::Capability UpdateManager::capabilityFromString(const QString& name) {
